Rename module from `pijul` to `pijul_channel`.
