+++ src/modules/mod.rs
@@ -58,7 +58,7 @@ mod os;
 mod package;
 mod perl;
 mod php;
-mod pijul;
+mod pijul_channel;
 mod pulumi;
 mod purescript;
 mod python;
@@ -160,7 +160,7 @@ pub fn handle<'a>(module: &str, context: &'a Context) -> Option<Module<'a>> {
             "package" => package::module(context),
             "perl" => perl::module(context),
             "php" => php::module(context),
-            "pijul" => pijul::module(context),
+            "pijul_channel" => pijul_channel::module(context),
             "pulumi" => pulumi::module(context),
             "purescript" => purescript::module(context),
             "python" => python::module(context),
@@ -275,7 +275,7 @@ pub fn description(module: &str) -> &'static str {
         "package" => "The package version of the current directory's project",
         "perl" => "The currently installed version of Perl",
         "php" => "The currently installed version of PHP",
-        "pijul" => "The current channel of the repo in the current directory",
+        "pijul_channel" => "The current channel of the repo in the current directory",
         "pulumi" => "The current username, stack, and installed version of Pulumi",
         "purescript" => "The currently installed version of PureScript",
         "python" => "The currently installed version of Python",
